# IELTSLISA_ZaloApp_User/IELTSLISA_ZaloApp_User/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "IELTSLISA_ZaloApp_User, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/Class/GetClassById": {
      "get": {
        "tags": [
          "Class"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "classId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Class"
            }
          }
        }
      }
    },
    "/Class/AddNewClass": {
      "post": {
        "tags": [
          "Class"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "className",
            "type": "string"
          },
          {
            "in": "query",
            "name": "classContent",
            "type": "string"
          },
          {
            "in": "query",
            "name": "classImg",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/Feedback/GetFeedbackById": {
      "get": {
        "tags": [
          "Feedback"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "feedbackId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Feedback"
              }
            }
          }
        }
      }
    },
    "/Feedback/AddNewFeedback": {
      "post": {
        "tags": [
          "Feedback"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "feedbackTile",
            "type": "string"
          },
          {
            "in": "query",
            "name": "feedbackContent",
            "type": "string"
          },
          {
            "in": "query",
            "name": "feedbackName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "feedbackClass",
            "type": "string"
          },
          {
            "in": "query",
            "name": "feedbackPic",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/Information/GetInformationById": {
      "get": {
        "tags": [
          "Information"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "infoId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Information"
              }
            }
          }
        }
      }
    },
    "/Information/AddNewInformation": {
      "post": {
        "tags": [
          "Information"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "infoName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "infoImg",
            "type": "string"
          },
          {
            "in": "query",
            "name": "infoContent",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/User/AddNewUser": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "userName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "phone",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/User/GetUserByPhone": {
      "get": {
        "tags": [
          "User"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "phone",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
...
    "/Voucher/GetNotOwnVoucher": {
      "get": {
        "tags": [
          "Voucher"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Voucher"
              }
            }
          }
        }
      }
    },
    "/Voucher/AddNewVoucher": {
      "post": {
        "tags": [
          "Voucher"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "voucherCode",
            "type": "string"
          },
          {
            "in": "query",
            "name": "voucherName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "voucherDescription",
            "type": "string"
          },
          {
            "in": "query",
            "name": "startDate",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "endDate",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Voucher/UpdateVoucher": {
      "put": {
        "tags": [
          "Voucher"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "voucherId",
            "type": "string"
          },
          {
            "in": "query",
            "name": "voucherCode",
            "type": "string"
          },
          {
            "in": "query",
            "name": "voucherName",
            "type": "string"
          },
          {
            "in": "query",
            "name": "voucherDescription",
            "type": "string"
          },
          {
            "in": "query",
            "name": "startDate",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "endDate",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Voucher/GetVoucherById": {
      "get": {
        "tags": [
          "Voucher"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "voucherId",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Voucher"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Class": {
      "type": "object",
      "properties": {
        "classId": {
          "type": "string"
        },
        "className": {
          "type": "string"
        },
        "classContent": {
          "type": "string"
        },
        "classImg": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Feedback": {
      "type": "object",
      "properties": {
        "fbId": {
          "type": "string"
        },
        "fbTitle": {
          "type": "string"
        },
        "fbContent": {
          "type": "string"
        },
        "fbName": {
          "type": "string"
        },
        "fbClass": {
          "type": "string"
        },
        "fbPic": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Gift": {
      "type": "object",
      "properties": {
        "giftId": {
          "type": "string"
        },
        "giftName": {
          "type": "string"
        },
        "giftDescription": {
          "type": "string"
        },
        "giftQuantity": {
          "format": "int32",
          "type": "integer"
        },
        "giftStatus": {
          "type": "boolean"
        },
        "userVouchers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserVoucher"
          }
        },
        "voucherGifts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/VoucherGift"
          }
        }
      },
      "additionalProperties": false
    },
    "Information": {
      "type": "object",
      "properties": {
        "infoId": {
          "type": "string"
        },
        "infoName": {
          "type": "string"
        },
        "infoImg": {
          "type": "string"
        },
        "infoContent": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "User": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "userVouchers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserVoucher"
          }
        }
      },
      "additionalProperties": false
    },
    "UserVoucher": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "voucherId": {
          "type": "string"
        },
        "userVoucherStatus": {
          "type": "boolean"
        },
        "redeemed": {
          "format": "date-time",
          "type": "string"
        },
        "giftId": {
          "type": "string"
        },
        "gift": {
          "$ref": "#/definitions/Gift"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "voucher": {
          "$ref": "#/definitions/Voucher"
        }
      },
      "additionalProperties": false
    },
    "Voucher": {
      "type": "object",
      "properties": {
        "voucherId": {
          "type": "string"
        },
        "voucherCode": {
          "type": "string"
        },
        "voucherName": {
          "type": "string"
        },
        "voucherDescription": {
          "type": "string"
        },
        "startDate": {
          "format": "date-time",
          "type": "string"
        },
        "endDate": {
          "format": "date-time",
          "type": "string"
        },
        "voucherStatus": {
          "type": "boolean"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "userVouchers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserVoucher"
          }
        },
        "voucherGifts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/VoucherGift"
          }
        }
      },
      "additionalProperties": false
    },
    "VoucherGift": {
      "type": "object",
      "properties": {
        "voucherId": {
          "type": "string"
        },
        "giftId": {
          "type": "string"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "gift": {
          "$ref": "#/definitions/Gift"
        },
        "voucher": {
          "$ref": "#/definitions/Voucher"
        }
      },
      "additionalProperties": false
    },
...
  }
}